requirements.txt
asgiref==3.7.2
asttokens==2.4.0
backcall==0.2.0
black==23.10.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
cloudinary==1.36.0
colorama==0.4.6
cryptography==41.0.4
decorator==5.1.1
defusedxml==0.8.0rc2
Django==4.2.6
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djoser==2.2.0
executing==2.0.0
idna==3.4
ipython==8.16.1
isort==5.12.0
jedi==0.19.1
matplotlib-inline==0.1.6
mypy-extensions==1.0.0
mysqlclient==2.2.0
oauthlib==3.2.2
packaging==23.2
parso==0.8.3
pathspec==0.11.2
pickleshare==0.7.5
platformdirs==3.11.0
prompt-toolkit==3.0.39
pure-eval==0.2.2
pycparser==2.21
Pygments==2.16.1
PyJWT==2.8.0
python-decouple==3.8
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.4.2
sqlparse==0.4.4
stack-data==0.6.3
traitlets==5.11.2
tzdata==2023.3
urllib3==2.0.7
wcwidth==0.2.8
